missed int64 -> int reversion
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@9378 a380766d-ff14-0410-b294-a243070f3f08

+++ Headers/server/SC_ComPort.h
@@ -58,7 +58,7 @@ class SC_ComPort : public SC_CmdPort
 {
 protected:
 	int mPortNum;
-	int64 mSocket;
+	int mSocket;
 	struct sockaddr_in mBindSockAddr;
 
 #ifdef USE_RENDEZVOUS
